Checkpoint gameplay flow and results decompilation
Preserve the pending gameplay, results-screen, ABI and host-audio work after the squash merge. Reconstruct the two-pass gizmo flow parser and its constructors from the original binary; flow reset/execution and results initialization remain unfinished. This is a buildable development checkpoint, not a claim of completed gameplay fixes.

Target, native and WASM builds pass, as do all four repository checks. Regenerate the matching report. Overall fuzzy matching rises from 25.23344% to 25.69756%, with 1257 exact functions and no lost exact matches. Partial functions still have regressions, including MenuText3DEx (82.578316% to 69.445786%) and incomplete results/flow functions; these require further matching work. Required checks ran manually; skip the optional whole-worktree formatting hook.

+++ src/legoapi/render/core/render.h
@@ -16,7 +16,7 @@ void Draw_AUTOSAVEWARNING(void);
 #ifdef __cplusplus
 extern "C" {
 #endif
-    void Text3DEx(char *, f32, f32, f32, f32, f32, f32, u8, u8, u8, u8, i32);
+    void Text3DEx(char *, f32, f32, f32, f32, f32, f32, u32, u8, u8, u8, i32);
 #ifdef __cplusplus
 }
 #endif
